# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Overthrone game API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterTeamBody {
  name: string;
  password: string;
  members?: string[];
}

export interface LoginTeamBody {
  name: string;
  password: string;
}

export interface Team {
  id: number;
  name: string;
  hp: number;
  ap: number;
  members: string[];
  isEliminated: boolean;
  /** @nullable */
  allianceId: number | null;
  /** @nullable */
  activeTaskId: number | null;
  isAdmin: boolean;
  createdAt: string;
}

export interface AuthResponse {
  team: Team;
  token: string;
}

export type GameStateStatus =
  (typeof GameStateStatus)[keyof typeof GameStateStatus];

export const GameStateStatus = {
  waiting: "waiting",
  active: "active",
  attack_phase: "attack_phase",
  ended: "ended",
} as const;

export type GameStatePhase =
  (typeof GameStatePhase)[keyof typeof GameStatePhase];

export const GameStatePhase = {
  task: "task",
  attack: "attack",
} as const;

export interface GameState {
  status: GameStateStatus;
  currentEpoch: number;
  /** @nullable */
  epochStartedAt: string | null;
  /** @nullable */
  epochEndsAt: string | null;
  epochDurationSeconds: number;
  totalEpochs: number;
  phase: GameStatePhase;
}

export type TaskType = (typeof TaskType)[keyof typeof TaskType];

export const TaskType = {
  sudoku: "sudoku",
  math: "math",
  ctf: "ctf",
  algorithm: "algorithm",
} as const;

export type TaskDifficulty =
  (typeof TaskDifficulty)[keyof typeof TaskDifficulty];

export const TaskDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export interface Task {
  id: number;
  title: string;
  description: string;
  type: TaskType;
  difficulty: TaskDifficulty;
  apReward: number;
  content: string;
  isActive: boolean;
}

export interface SubmitTaskBody {
  answer: string;
}

export interface TaskSubmissionResult {
  correct: boolean;
  apEarned: number;
  message: string;
}

export interface UseTaskCardBody {
  taskId: number;
}

export interface UseAttackCardBody {
  targetTeamId: number;
  apSpent: number;
  useAI?: boolean;
}

export interface UseAllianceCardBody {
  targetTeamId: number;
}

export interface AllianceResponseBody {
  requestingTeamId: number;
  accept: boolean;
}

export interface CardActionResult {
  success: boolean;
  message: string;
  teamState?: Team;
}

export interface AttackResult {
  success: boolean;
  message: string;
  damageDealt: number;
  apSpent: number;
  targetTeamId: number;
  targetTeamName: string;
}

export interface SuspicionResult {
  correct: boolean;
  message: string;
  accusingTeamEliminated: boolean;
  accusedTeamEliminated: boolean;
}

export interface LeaderboardEntry {
  rank: number;
  teamId: number;
  teamName: string;
  hp: number;
  ap: number;
  totalScore: number;
  isEliminated: boolean;
  /** @nullable */
  allianceId: number | null;
  /** @nullable */
  allianceName: string | null;
  /** @nullable */
  allianceHP: number | null;
  /** @nullable */
  allianceAP: number | null;
}

export interface Leaderboard {
  entries: LeaderboardEntry[];
  updatedAt: string;
}

export interface KingdomMapEntry {
  teamId: number;
  teamName: string;
  hp: number;
  ap: number;
  isEliminated: boolean;
  /** @nullable */
  allianceId: number | null;
  x: number;
  y: number;
  size: number;
}

export type GameEventType = (typeof GameEventType)[keyof typeof GameEventType];

export const GameEventType = {
  attack: "attack",
  alliance_formed: "alliance_formed",
  alliance_broken: "alliance_broken",
  backstab: "backstab",
  task_completed: "task_completed",
  team_eliminated: "team_eliminated",
  suspicion: "suspicion",
} as const;

export interface GameEvent {
  id: number;
  type: GameEventType;
  /** @nullable */
  fromTeamId: number | null;
  /** @nullable */
  fromTeamName: string | null;
  /** @nullable */
  toTeamId: number | null;
  /** @nullable */
  toTeamName: string | null;
  description: string;
  epoch: number;
  createdAt: string;
}

export interface AdminTeamDetail {
  id: number;
  name: string;
  hp: number;
  ap: number;
  members: string[];
  isEliminated: boolean;
  /** @nullable */
  allianceId: number | null;
  /** @nullable */
  activeTaskId: number | null;
  isAdmin: boolean;
  tasksCompleted: number;
  attacksMade: number;
  createdAt: string;
}

export type CreateTaskBodyType =
  (typeof CreateTaskBodyType)[keyof typeof CreateTaskBodyType];

export const CreateTaskBodyType = {
  sudoku: "sudoku",
  math: "math",
  ctf: "ctf",
  algorithm: "algorithm",
} as const;

export type CreateTaskBodyDifficulty =
  (typeof CreateTaskBodyDifficulty)[keyof typeof CreateTaskBodyDifficulty];

export const CreateTaskBodyDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export interface CreateTaskBody {
  title: string;
  description: string;
  type: CreateTaskBodyType;
  difficulty: CreateTaskBodyDifficulty;
  content: string;
  answer: string;
}

export type ListTasksParams = {
  difficulty?: ListTasksDifficulty;
};

export type ListTasksDifficulty =
  (typeof ListTasksDifficulty)[keyof typeof ListTasksDifficulty];

export const ListTasksDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;