# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Overthrone game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerTeam
      tags: [auth]
      summary: Register a new team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterTeamBody"
      responses:
        "201":
          description: Team registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: loginTeam
      tags: [auth]
      summary: Login as a team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginTeamBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current logged in team
      responses:
        "200":
          description: Current team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "401":
          description: Unauthorized
...
  /admin/teams/{id}/eliminate:
    post:
      operationId: eliminateTeam
      tags: [admin]
      summary: Admin - eliminate a team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team eliminated
  /admin/tasks:
    post:
      operationId: createTask
      tags: [admin]
      summary: Admin - create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
components:
  schemas:
...
    RegisterTeamBody:
      type: object
      properties:
        name:
          type: string
        password:
          type: string
        members:
          type: array
          items:
            type: string
      required:
        - name
        - password
    LoginTeamBody:
      type: object
      properties:
        name:
          type: string
        password:
          type: string
      required:
        - name
        - password
    AuthResponse:
      type: object
      properties:
        team:
          $ref: "#/components/schemas/Team"
        token:
          type: string
      required:
        - team
        - token
    Team:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        hp:
          type: integer
        ap:
          type: integer
        members:
          type: array
          items:
            type: string
        isEliminated:
          type: boolean
        allianceId:
          type: ["integer", "null"]
        activeTaskId:
          type: ["integer", "null"]
        isAdmin:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - hp
        - ap
        - members
        - isEliminated
        - allianceId
        - activeTaskId
        - isAdmin
        - createdAt
...
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [sudoku, math, ctf, algorithm]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        apReward:
          type: integer
        content:
          type: string
        isActive:
          type: boolean
      required:
        - id
        - title
        - description
        - type
        - difficulty
        - apReward
        - content
        - isActive
...
    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [sudoku, math, ctf, algorithm]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        content:
          type: string
        answer:
          type: string
      required:
        - title
        - description
        - type
        - difficulty
        - content
        - answer